move app code to app folder
